Split out connection mocks
The network mocks header is quite large. I only need access to the
connection mocks for my UT. So, split out the network mocks into their
own header to decrease test compile times.
