+++ test/mocks/network/BUILD
@@ -9,10 +9,22 @@ load(
 envoy_package()
 
 envoy_cc_mock(
+    name = "connection_mocks",
+    srcs = ["connection.cc"],
+    hdrs = ["connection.h"],
+    deps = [
+        "//include/envoy/network:connection_interface",
+        "//test/mocks/event:event_mocks",
+        "//test/mocks/stream_info:stream_info_mocks",
+    ],
+)
+
+envoy_cc_mock(
     name = "network_mocks",
     srcs = ["mocks.cc"],
     hdrs = ["mocks.h"],
     deps = [
+        ":connection_mocks",
         "//include/envoy/buffer:buffer_interface",
         "//include/envoy/network:connection_interface",
         "//include/envoy/network:drain_decision_interface",
